contrib-modules, passwd/totp: Merge authTimestamp from lastbind overlay.
This code duplicates the basic function of lastbind. The two overlays
cannot be used together. The timestamp Mod op is changed to require
the old value to still be present at the end of the Bind. This allows
us to detect collisions (multiple successful Binds in the same time
window) and properly fail the extra Bind attempts.

Change-Id: Ib29ef8e8ab9ac77d2004b54bf8a7c8b85b2545d9