tests: fix sporadic failures of tests related to syncrepl.
Explicit specify 'retry="1 +" for all syncrepl tests.
Otherwise the default retry interval will be used,
which is about 60 seconds. In such cases the first
syncrepl attempt could fail.

For example, when a provider was empty on start and
was not in time get a data from its own parent-provider.

So the first attempt will fail at the remote-search phase,
syncrepl will sleep for a 60 secondsm,
and we will get test failure by timeout.

Change-Id: Iff4f3983c3eca09e349c8c29c039a7e8b10218d1