syncrepl: use biglock to avoid troubles with uuid-2-dn mapping.
This is 1-of-5 patch for https://github.com/ReOpen/ReOpenLDAP/issues/43

Replication receives notifications and list of entries in as
UUID. But any updates could be applies to DIT only by using DN
as a primary key. Therefore syncrepl performs translation from
UUID to DN by searching.

But such uuid2dn mapping could be changed asynchronously by
other threads. For instance, a 'moddrn' request changes it
explicitly. Also a 'delete' request with immediately following
'add' creates a new DN-entry with a new UUID.

To allow replication work properly uuid2dn mapping must not be
changes while syncrepl does something. Therefore at least
'moddrn' and 'add' operations should be synchronized with
syncrepl to avoid volatility of uuid-2-dn mapping.

However, in a multi-node multi-master cluster notifications
comes from all nodes, and therefore could be reordered relative
to initial sequence of changes. Therefore 'modify' operation
should also be synthesized to avoid overwriting the last (but
latecomer) update by a one of previous.

This patch synchronizes ALL WRITE operations, including
'delete', with the syncrepl's message loop. So, this is a bit
excessive that minimum necessary. But it is more safe and just
a workaround for other bugs (which are exists, sure). Also it
keeps implementation of the 'biglock' simple, clear and
reliable.

Change-Id: I219a5504da09ef9bf346d8d2d425e4d7c7943913