syncrepl: fix flaw in `delete-non-present`.
This is 2-of-5 patch for https://github.com/ReOpen/ReOpenLDAP/issues/43

Original implementation of delete-non-present, has a big flaw
in the logic.

When delete-non-present selects an entries for deletion, it
should check that entryCSN is NOT recent than the CSN of the
same SID in the cookie from a provider.

Otherwise it will delete the entries with created after the
provider was synchronized with the origin, where such entries
had created.

This patch adds a such checking.

Change-Id: I59ea5904208b8286b957408d9e52708fc85c0c38