libreldap: MozNSS load certificates from certdb, fallback to PEM.
Base on http://pkgs.fedoraproject.org/cgit/rpms/openldap.git/plain/openldap-nss-certs-from-certdb-fallback-pem.patch

If TLS_CACERT pointed to a PEM file and TLS_CACERTDIR was set to NSS
certificate database, the backend assumed that the certificate is always
located in the certificate database. This assumption might be wrong.

This patch makes the library to try to load the certificate from NSS
database and fallback to PEM file if unsuccessfull.

Upstream ITS: #7389
Resolves: RHEL's #857455

Change-Id: I045ef68d6bb6dbc49b7f7e17048b6020b51d21f8