libreldap: remove resolv-mutex around getnameinfo() and getnameinfo() (Debian#340601).
OpenLDAP upstream conservatively assumes that certain resolver functions
(getaddrinfo, getnameinfo, res_query, dn_expand) are not re-entrant; but we
know that the glibc implementations of these functions are thread-safe, so
we should bypass the use of this mutex.  This fixes a locking problem when
an application uses libldap and libnss-ldap is also used for hosts
resolution.

Closes Debian bug #340601.

Change-Id: I6b9aa11f1c2bf51b0adc8c0f646d3f053761fecf