backend-bdb: don't second-guess BDB ABI (Debian#651333).
OpenLDAP upstream conservatively assumes that any change to the
version number of libdb can result in an API-breaking change
that could impact the database.

In Debian (and other modern distros), we know that such changes
require bumping the library soname and changing the package
name, and demand such rigor from our package maintainers even
when upstreams don't deliver; so any such check in the source
code works against the packaging system by forcing database
upgrades when we know none are required. Disable this check so
we rely on the packaging system to do its job.

Change-Id: I4355b0a458d7dc1163e7d44d00e7f7b604402906
Bug-Debian: http://bugs.debian.org/651333