contrib: argon2 password hashing module (ITS#8575).
This change implements argon2, which won the Password Hashing
Competition (https://password-hashing.net/) as a contrib-module in order
to provide a modern password hashing alternative in openldap. The
currently available password hashing algorithms are relatively old, and
modern hardware, especially GPUs can compute quite a few (ranging from
tens of thousands to millions) of hashes per second. Argon2 was designed
to withstand such attacks.

To implement the argon2 function, this module links against libsodium,
which exposes an API for password hashing that uses the argon2 function.

This implementation uses the default work factors used in the libsodium
implementation for "interactive" use.

Change-Id: Ib8994d2bdcd50aea5c40d0ae6fb81c061704c693