libreldap, slapd: backport - Fixes for multiple threadpool queues.
Remove poolq_hash, it wasn't distributing work evenly to the queues.
Just walk through all queues and use the one with smallest
active+pending count. Since pool_retract also relied on the hash,
a different means of locating the thread to retract was needed.
Add pool_submit2 which returns the threadpool task structure,
and record which poolq this task lives on.

Change-Id: I7f11f98c9926cbf6528a757ab6dc85d9b8a6d414